Avoid downloading base packages if possible. Fixes #345.
Before manually extracting a sub package from a base package, the package description of the base package is first searched for the sub package. This avoids the otherwise necessary download of the base package whenever the sub package is a non-path-based one. This also opens the way for the dub registry to always collect all sub packages into the reported package description, so that it will never be necessary to actually download a package for getting a sub package.