Get .visualdproj files to be created in the .dub directory. Less cluttering of the working directory.
However this creates a problem since when debugging an executable through VS the default working directory will be .dub/ where it was ./ previously. This can cause eg. files not being found.

This "debugging working directory" setting is stored in the .suo file generated by the IDE, so maybe a VisualD option is needed to get out of this.