Remove POSIX special-casing from copyFile
Since `std.file.copy` provides a means to preserve attributes of a file
as of DMD 2.067.0 or later, it is preferable to use this option rather
than having a custom-written solution.  `version (Posix)` special-casing
is therefore used only where the DMD version does not support it.

Note that where the `version (Posix)` code uses a `chown` call followed
by a `chmod` call, `std.file.copy` only uses a `chmod`.  This is not
expected to cause any issues but should be borne in mind as a possible
breaking change for some use-cases.

Fixes https://github.com/dlang/dub/issues/992.