Fix #1686: Running single-file package as pre generate command always rebuilds
If a single-file Dub package is used as a pre generate command the
project containing the pre generate command will always be rebuilt.
This happens because the path to the single-file package is not
correctly added to the package. This will later lead to the current
directory, `.`, being added as a dependency instead of the path to
single-file package.