make selected nonmatching dependencies abort build
We had this as discussion at dconf. Changing the dependency version
inside dub.json/sdl should fail the build if the fetched dependency is
unmatching. This will break previous "hacks" where dub.selections.json
was manually edited to override the version.

dub.json should be considered as what the developer wanted and therefore
should not be overwritten with just a warning message, but instead abort
the current build.