Fix issue 1277: Missing dub.json causes infinite loop
For some reason, having a 'source/app.d' was recognized as being in a package.
This would load an empty package with no manifest and send the dependency resolver
in an infinite loop.
Remove the special detection of 'source/app.d' and rely only on manifest.
Another option would have been to detect 'source/app.d' as a single-file package,
but that is a bit too much magic (the user can just 'dub run --single source/app.d').