Move most CI scripts in scripts/ci
The root directory was getting a bit crowded.
semaphore-ci.sh was not moved, because there's no visible way to configure semaphore from the repository.
Additionally, build.d is left in the root directory as it's useful to have it here,
and it'll be only one script one build.sh, build.cmd, and build-gdc.sh are gone.