Raise minimum Vibe.d requirement to v0.9.0
Make sure a recent enough Vibe.d is used so we can update the vibecompat code.
Additionally, since the split of Vibe.d into package, and the splitting of
vibe-core into its own project, the version specification was never triggered,
so it was corrected in this commit (the version spec was correct for older Vibe.d).