Adapt determinePlatform to the new defaults for Win32
Starting with DMD 2.099, `-m32` defaults to MsCOFF instead of OMF and
`-m32omf` is required to use the OMF format + OPTLINK.

This commit changes the logic to select the appropriate logic based
on the current compiler version. Potentially using the wrong object file
format during platform probing is fine because the codegen is skipped
(dmd is invoked with `-c -o-`).