Move fix for issue 711 outside of JSON/SDL parser
Having the fix for this issue in the parser means adding non-trivial logic.
It is easier to reason about to have this logic done as a post-processing step,
replacing the `autodetect` with the actually auto-detected version.
This also paves the way to have other `targetType` as default,
e.g. the `unittest` configuration should probably not be `library`.