Parse settings.json using configy
This will provide superior error messages to the user,
and match what we already do for `dub.selections.json`.

It also allows us to remove many fields from the `Dub` class.
Those fields existed so that we would not have to re-compute the data
from the JSON DubConfig was storing. Now that we use a simple struct,
and do the aggregation eagerly, all of that complexity can go.