Overrides: Make it clear that only VersionRange can be overriden
The code was always assuming overrides were VersionRange,
as trying to do 'dub add-override' would call 'Dependency'
constructor with a string, which forwards to VersionRange.fromString.
However, when using dub as a library, this assumption was much less obvious,
notably because the fields are public.
With this change, we make sure library users do not misuse overrides.