Remove `Dub.updatePackageSearchPath`
This function was essentially two functions: One when an override was set,
another one for the common case. However, the function is private,
and the number of call sites is limited, so it's better to inline
the call at the caller site. Once that was done, it became obvious
it only ever has an effect once, when the Dub instance is constructed,
as we don't modify this environment variable.
Thus, we can inline it in the constructor, and remove the dead branches.