Integrate `dependencyBuildSettings` into dependencies
Because those are specified in `dependencies`,
they should be stored alongside them.
Unfortunately this will slihtly increase memory usage in the common case,
but unlikely to a point where it would be problematic.

This is technically a breaking change as a field is getting removed,
however this struct has been manipulated a few times in breaking ways,
and it's unlikely `dependencyBuildSettings` is being used outside of dub,
given how recent it is and how narrow its usage is.
However, as `dependencies` is likely to be widely used, a compatibility
type and an `alias this` are provided not to break client code.