PackageManager: Load packages lazily on dub build
A problem of the current design of dub is that all locations are scanned,
regardless of what action is being performed, often multiple times.
With this change, a regular `dub build` on a project with all dependencies
resolved will no longer loads all available package, which can substantially
speed up development.