PackageManager: Move the scanning logic to Location, segment per Location
Before this change, all packages were surprisingly stored in m_internal.localPackages,
which itself is a new addition to replace an array of 'Package'.
This means that even though we had an architecture that looks like it segmented
things per repository / location, it actually didn't,
and things worked solely based on the order packages were added it
(and the fact we did a linear search).