PackageManager.Location: Split packages registered locally and read from path
A Location, having a default search path (the package path),
and some extra search path, optionally, previously stored
all of them in the same array. However, this creates problems,
as we want to move to lazy load the packages in the default
search path, while we can't lazy load packages in non-default ones.