PackMan: Deprecate overload leaking location path
Currently, a lot of the internals of the PackageManager are leaking.
The getPackage overload accepting both a version and a NativePath
is an example of such a leak, and looking at the usage sites
show that things can be simplified by taking a PlacementLocation,
which is a higher level representation of the path.