PackageManager: Only iterate on internal fromPath array, not everything
getOrLoadPackage is only used for path-based dependencies,
which are only stored in m_internal.fromPath.
We can thus reduce the iteration by only using fromPath.
If someone had a path-based dependency that pointed to a version-based package,
it would lead to this package being loaded twice, but that's not really a concern,
as it would not lead to problem either.