Do not fall back to Vibe.d in vibecompat
This remove the fallback to Vibe.d, as there is no plan in sight
to remove the compatibility layer, hence we always end up depending
on something out of date, or rely on a contributor upgrading the code.
In other words, having the fallback complicates development.

As a result, this removes:
- The outdated, now empty `dub.selections.json`;
- The code in `Have` versions;
- Some old methods that import vibe and couldn't possibly compile;