Replace `settings.json` with `settings.yaml` and provide a new syntax
This adds parsing for `settings.yaml` (`settings.yml` is not supported),
which takes precedence over `settings.json`.
As we are introducing a new feature, this also slightly change the format
to be more intuitive, consistent, and less repetitive.
The main change is on the way one can specify registries: instead of having
two fields, only have one, which might lead to a bit more repetition in
some cases, but generally should be more intuitive to users.