remove GeneratorSettings.cache
instead the cache is a property of PackageManager

This way dub lib users don't have to try to figure out where the real
cache path is and can just continue using the generator and describe
APIs like in previous versions.

Customizing cache path is no longer easily possible and now must be done
through creating a custom PackageManager and use that on `new Project()`
to pass into the generators. If needed, APIs to modify m_dirs on
startup could be added to Dub subclasses in the future.