packagesuppliers/filesystem.d: Only look for versions of package_id
When looking for versions of a package don't include any files that
aren't versions of the queried package.

If a directory contains the files botan-1.zip and botan-math-2.zip
the current implementation would report both 1 and math-2 as
candidates followed by an error that math-2 is not a valid version.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>