Change PackageSupplier interface to accept VersionRange
This is a leftover of the time when everything in dub accepted Dependency.
While the mess had been untangled, doing so on this interface proved harder,
as there is no good way to deprecate it in a fully backward-compatible manner
without introducing a lot of overhead (a new type). Instead, this uses
a middle ground approach, where the external implementers (which most likely
do not exists) would break (although the fix is trivial), but callers would not.