Remove fallback to ~master on dub fetch, fix #2679
The retry does not make sense, as one cannot add a package
to the registry without a release version.
It is also not in line with Dub strategy, as we are trying
to reduce the usage of branch dependencies.

This also fixes issue #2679 which pointed out the error message
was sub-par. The API of the registry seems to have changed and
we now get `null` instead of `undefined`, leading to an error
inside the JSON module instead of the enforce being triggered.