Deprecate Project.packageManager
There is no reason to expose the PackageManager through the Project.
In fact, the only two places it was used were in generators,
where the PackageManager was stored but never used.
The generators arguably should not have access to the PackageManager,
as all the Packages should have been loaded and handled to them
before the generation starts.