Adapt PackageSuppliers to PackageName API
Since we recently did a semi-breaking change to PackageSuppliers,
now is the time to do it again (they will be in the same release).
Again the same logic applies: it's only a breaking change for those
inheriting from PackageSupplier, not those calling it,
and there is likely not a lot of people doing either out there.