Tests: Use TestDub FSEntry constructor for more accurate tests
The previous approach added Packages in the PackageManager directly,
completely sidestepping the scanning the PackageManager does,
which is a behavior we absolutely want to test.

By setting up the filesystem before we instantiate our `TestDub` instance,
we can remove a bunch of exceptions / overrides, and get a much closer
behavior to the final product.