Return early from determineDependencyConfigs if possible
The only package that can have a config already created is the root
package, and only if the `config` parameter isn't an empty string.
Every other package is in the `configs` array iff it was added by
a previous call to this function. The exact same configurations
will be found if this function is called again for the same (p, c)
pair and it will do needless work that won't modify anything.