Respect skipRegistry value from configuration files
Currently, the skipRegistry value in configuration files is ignored,
the final value is either the one the user passed on the command line,
which is the intended behavior, or the value `none` if the user didn't
pass a --skip-registry argument.

This commit fixes this by introducing the new value `default_` to the
SkippackageSuppliers enum to know when the user didn't pass the
skip-registry argument and the value should be taken from the
configuration files.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>