feat(selections): Add the ability to read integrity tag from selections
The `dub.selections.json` file can now contains integrity tags matching the
SRI specifications, allowing dub (and other tools, e.g. Nix) to better validate
that the downloaded archive matches the expected version.

However, Dub will not yet write the integrity tag, as it would result in a bad
user experience. Since `dub` tries hard to reuse packages present on the
filesystem, doing a `dub upgrade` could wipe the integrity tag (or not populate it)
if the package is already present on the system, an issue which would manifest itself
quite often for popular packages.

In order to solve this issue, we could store the integrity tag on disk,
however this can be done in another PR as such package metadata would be
useful for other purposes as well.