fix(mockfs): Improve Windows support and add tests
On Windows, the root has a name, which is the drive,
unlike on POSIX where the root is an empty name.
This led to a flurry of problems because we assumed
a POSIX style, where the FS is a directed graph,
while on Windows it is multiple unconnected graphs
(the drives cannot be accessed via a relative path).