packagesuppliers/filesystem.d: Fix out of bounds on invalid pkg names
If a user ran `dub fetch foo-1.1.1` instead of `dub fetch foo@1.1.1`
over a filesystem package supplier, assuming that the archive exists,
the code that slices the filename tries to create an invalid slice as
it assumes that the filename will contain at least the package name, a
one character version separator, and the archive suffix which wouldn't
be the case with a package name like `foo-1.1.1` and an archive name
of `foo-1.1.1.zip`.

The fix is to require one more character in the glob to account for
the version separator.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>