feat: Implement support for Github-based index, bypassing the registry
This implement a package 'index' similar to that found for Homebrew, Nix, Cargo, etc...
It allows us to remove a SPOF in our critical infrastructure, as a Github outage
would always cause a registry being unusable anyway.

There are multiple steps to having a useful index:
- For transition purpose, we add a hidden command to Dub that export an `index.yaml`;
- In the future, users should register their packages by adding an entry to `index.yaml`,
  the index definition file of the registry. This is used as the source of all packages;
- `dub` now has a hidden `index-build` command to allow it to build the index based
  on an index definition file (`index.yaml`). Using this, it queries the various APIs
  to generate JSON index files that are stored under a pre-defined hierarchy.
- Finally, a `PackageSupplier` is added to make use of this new feature;

In the future, the registration process needs to be moved from the registry to Github
to make this migration complete. This *can* be done by exposing a user-friendly interface
on `code.dlang.org`, if making an MR to the index is deemed too complicated.