feat(recipe): Add support YAML recipe
This makes dub recognize `dub.yaml` as a configuration file format.
The code is very similar to the JSON one, as Dub has been using the
YAML parser for multiple years. It allows us to use a configuration
format that is in widespread use and is more human-oriented than JSON.