fix(main): Re-enable the default D GC
This was disabled a few years ago as a workaround for a bug that was
mostly affecting Dub running on Travis CI. At the time, the osthreads
of DRuntime didn't take TLS into account, and due to how pthread works,
the TLS was eating away at what we thought was available stack to the thread.
Since this is long fixed and we don't support such old versions,
we can remove this.